refactor cachedicbcadapter client
